feat: share base utils through a newly created shared package
