+++ packages/config/tsconfig-react.json
@@ -1,0 +1,8 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "display": "React Library",
+  "extends": "./tsconfig-base.json",
+  "compilerOptions": {
+    "jsx": "react-jsx"
+  }
+}
