+++ packages/magickbase/next.config.mjs
@@ -3,6 +3,9 @@ const i18nConfig = (await import('./next-i18next.config.js')).default
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   i18n: i18nConfig.i18n,
+
+  transpilePackages: ['@magickbase-website/shared'],
+
   webpack(config) {
     config.module.rules.push({
       test: /\.svg$/i,
