+++ packages/magickbase/package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@magickbase-website/shared": "workspace:^",
     "@radix-ui/react-dialog": "1.0.4",
     "@radix-ui/react-tooltip": "1.0.6",
     "@splinetool/react-spline": "2.2.6",
