+++ packages/neuron/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@docsearch/css": "3",
     "@docsearch/react": "3",
+    "@magickbase-website/shared": "workspace:^",
     "@octokit/graphql-schema": "^14.12.0",
     "@octokit/openapi-types": "^17.2.0",
     "@octokit/plugin-paginate-rest": "^6.1.2",
