+++ packages/neuron/src/pages/changelog/[[...slug]].page.tsx
@@ -6,7 +6,8 @@ import Link from 'next/link'
 import Image from 'next/image'
 import clsx from 'clsx'
 import { Fragment, useCallback } from 'react'
-import { Release, getReleases, range } from '../../utils'
+import { range } from '@magickbase-website/shared'
+import { Release, getReleases } from '../../utils'
 import { Page } from '../../components/Page'
 import styles from './index.module.scss'
 import ImgNeuronLogo from './neuron-logo.png'
