--- packages/neuron/src/utils/time.ts
@@ -1,1 +1,0 @@
-export const getTimeFormatter = (locale = 'en-US') => new Intl.DateTimeFormat(locale, { dateStyle: 'long' })
