+++ packages/shared/.eslintrc.cjs
@@ -1,0 +1,7 @@
+/** @type {import("eslint").Linter.Config} */
+const config = {
+  root: true,
+  extends: ['@magickbase-website/eslint-config/next.js'],
+}
+
+module.exports = config
