+++ packages/shared/.lintstagedrc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  '*.{js,cjs,mjs,jsx,ts,tsx}': ['prettier --write', 'eslint --fix'],
+  '*.{css,scss}': ['prettier --write', 'stylelint --fix'],
+}
