+++ packages/shared/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@magickbase-website/shared",
+  "version": "0.1.0",
+  "private": true,
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "typesVersions//": "This configuration is used to address the issue where projects depending on this package cannot correctly resolve the package path when using the Node moduleResolution mode. If the parent package uses the Bundler mode, this configuration may not be necessary. Refs: https://github.com/vercel/turbo/issues/4771#issuecomment-1587351163",
+  "typesVersions": {
+    "*": {
+      "index": [
+        "./src/index.ts"
+      ]
+    }
+  },
+  "devDependencies": {
+    "eslint": "^8.56.0",
+    "stylelint": "^15.6.2",
+    "stylelint-config-css-modules": "^4.2.0",
+    "stylelint-config-prettier-scss": "^1.0.0",
+    "stylelint-config-rational-order": "^0.1.2",
+    "stylelint-config-standard-scss": "^9.0.0",
+    "stylelint-order": "^6.0.3",
+    "typescript": "^5.1.3"
+  }
+}
