+++ packages/shared/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "@magickbase-website/config/tsconfig-nextjs.json",
+  "include": ["**/*.ts", "**/*.tsx"],
+  "exclude": ["node_modules"]
+}
